requirements.txt
aiohttp==3.8.4
aiohttp-retry==2.8.3
aiosignal==1.3.1
async-generator==1.10
async-timeout==4.0.2
attrs==23.1.0
blinker==1.6.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
docopt==0.6.2
exceptiongroup==1.1.1
execnet==1.9.0
Flask==2.3.2
frozenlist==1.3.3
greenlet==2.0.2
h11==0.14.0
idna==3.4
importlib-metadata==6.6.0
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
MarkupSafe==2.1.2
multidict==6.0.4
outcome==1.2.0
packaging==23.1
pluggy==1.0.0
protobuf==3.20.3
pycparser==2.21
PyJWT==2.6.0
PyMySQL==1.0.3
PySocks==1.7.1
pytest==7.3.1
pytest-xdist==3.2.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
regex==2023.5.5
requests==2.31.0
ruamel.yaml==0.17.26
ruamel.yaml.clib==0.2.7
selenium==4.9.1
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
tomli==2.0.1
tqdm==4.65.0
trio==0.22.0
trio-websocket==0.10.2
twilio==8.1.0
typing_extensions==4.6.3
urllib3==1.26.15
webdriver-manager==3.8.6
Werkzeug==2.3.3
wsproto==1.2.0
yarl==1.9.2
zipp==3.15.0
